check_tags_function/requirements.txt
google-cloud-bigquery
google-cloud-resource-manager
google-cloud-logging